requirements.txt
azure-functions==1.18.0
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.1.1
cryptography==41.0.7
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1